+++ packages/ckeditor5-dev-tests/lib/utils/automated-tests/getwebpackconfig.js
@@ -9,16 +9,24 @@ const path = require( 'path' );
 const escapedPathSep = path.sep == '/' ? '/' : '\\\\';
 const webpack = require( 'webpack' );
 const { getPostCssConfig } = require( '@ckeditor/ckeditor5-dev-utils' ).styles;
+const getDefinitionsFromFile = require( '../getdefinitionsfromfile' );
 
 /**
  * @param {Object} options
  * @returns {Object}
  */
 module.exports = function getWebpackConfigForAutomatedTests( options ) {
+	const definitions = Object.assign( {}, getDefinitionsFromFile( options.identityFile ) );
+
 	const config = {
 		mode: 'development',
 
+		watchOptions: {
+			aggregateTimeout: 500
+		},
+
 		plugins: [
+			new webpack.DefinePlugin( definitions ),
 			new webpack.ProvidePlugin( {
 				process: 'process/browser'
 			} )
@@ -77,6 +85,11 @@ module.exports = function getWebpackConfigForAutomatedTests( options ) {
 				'node_modules',
 				path.resolve( __dirname, '..', '..', '..', 'node_modules' )
 			]
+		},
+
+		output: {
+			// Get rid of the "webpack://" protocol to make the paths clickable in the terminal.
+			devtoolModuleFilenameTemplate: info => info.resourcePath
 		}
 	};
 
@@ -86,6 +99,12 @@ module.exports = function getWebpackConfigForAutomatedTests( options ) {
 		//
 		// See https://github.com/ckeditor/ckeditor5/issues/11006.
 		config.devtool = 'inline-source-map';
+
+		// Since webpack v5 it looks like splitting out the source code into the commons and runtime chunks broke the source map support.
+		config.optimization = {
+			runtimeChunk: false,
+			splitChunks: false
+		};
 	}
 
 	if ( options.coverage ) {