+++ packages/ckeditor5-dev-tests/tests/utils/automated-tests/assertions/equal-markup.js
@@ -26,7 +26,7 @@ describe( 'equalMarkup chai assertion', () => {
 			).to.equalMarkup(
 				'<paragraph>foo bXXX[]r baz</paragraph>'
 			);
-		} ).to.not.throw;
+		} ).to.not.throw();
 	} );
 
 	it( 'should throw AssertionError for unequal markups', () => {